Add files via upload
this is a simple script that locates the "claim allowance" button and auto clicks it. 
you must make sure the click location is set SPECIFICALLY for you (on the claim allowance button)
use an app like https://sourceforge.net/projects/mpos/ to find the coordinates and put the x value and z value in is appropriate position.